closes #225

+++ broker/src/main/java/pt/ulisboa/tecnico/softeng/broker/interfaces/BankInterface.java
@@ -4,8 +4,8 @@ import pt.ulisboa.tecnico.softeng.bank.dataobjects.BankOperationData;
 import pt.ulisboa.tecnico.softeng.bank.domain.Bank;
 
 public class BankInterface {
-	public static String processPayment(String IBAN, int amount) {
-		return Bank.processPayment(IBAN, amount);
+	public static String processPayment(String IBAN, double amount) {
+		return Bank.processPayment(IBAN,  amount);
 	}
 
 	public static String cancelPayment(String reference) {
